refactor: Extract data source application from routing
