+++ routers/expandRoutes.ts
@@ -1,29 +1,17 @@
-import * as _path from "node:path";
-import { applyUtilities } from "../htmlisp/utilities/applyUtility.ts";
-import { defaultUtilities } from "../htmlisp/defaultUtilities.ts";
 import { get } from "../utilities/functional.ts";
-import type {
-  Context,
-  DataSources,
-  Route,
-  Utilities,
-  Utility,
-} from "../types.ts";
+import type { DataSources, Route } from "../types.ts";
 
 async function expandRoutes({ routes, dataSources }: {
   routes: Record<string, Route>;
   dataSources: DataSources;
-}): Promise<
-  NonNullable<{ allRoutes: Record<string, Route>; allParameters: string[] }>
-> {
+}): Promise<Record<string, Route>> {
   const allRoutes: Record<string, Route> = {};
-  let allParameters: string[] = [];
 
   // Resolve promises in series to avoid race conditions in resolving
   for (
     const [url, route] of Object.entries(routes)
   ) {
-    const [expandedUrl, expandedRoute, parameters] = await expandRoute({
+    const [expandedUrl, expandedRoute] = await expandRoute({
       url,
       route,
       dataSources,
@@ -31,23 +19,19 @@ async function expandRoutes({ routes, dataSources }: {
     });
 
     allRoutes[expandedUrl] = expandedRoute;
-    allParameters = allParameters.concat(parameters);
   }
 
   // / is an exception. If it has an expansion, then it has to be added
   // to the root as otherwise the router won't find it later.
   if (allRoutes["/"]) {
     return {
-      allRoutes: {
-        ...allRoutes,
-        ...allRoutes["/"].routes,
-        "/": { ...allRoutes["/"], routes: {} },
-      },
-      allParameters,
+      ...allRoutes,
+      ...allRoutes["/"].routes,
+      "/": { ...allRoutes["/"], routes: {} },
     };
   }
 
-  return { allRoutes, allParameters };
+  return allRoutes;
 }
 
 async function expandRoute(
@@ -57,9 +41,8 @@ async function expandRoute(
     dataSources: DataSources;
     recurse: boolean;
   },
-): Promise<[string, Route, string[]]> {
+): Promise<[string, Route]> {
   let ret = { ...route };
-  let allParameters: string[] = [];
 
   if (route.expand) {
     const { matchBy } = route.expand;
@@ -87,9 +70,9 @@ async function expandRoute(
     );
 
     // Construct individual routes based on the results of indexing
-    const { slug, dataSources: matchByDataSources } = matchBy;
+    const { slug } = matchBy;
     if (Array.isArray(indexResults)) {
-      await Promise.all(indexResults.map(async (match) => {
+      await Promise.all(indexResults.map((match) => {
         const url = get(match, slug) as string;
 
         if (!url) {
@@ -100,35 +83,22 @@ async function expandRoute(
           );
         }
 
-        const { context, capturedParameters } = await getDataSourceContext(
-          Array.isArray(matchByDataSources)
-            ? matchByDataSources.map((dataSource) => ({
-              ...dataSource,
-              parameters: [match].concat(
-                dataSource.parameters,
-              ),
-            }))
-            : [],
-          dataSources,
-        );
-
-        if (capturedParameters) {
-          allParameters = allParameters.concat(capturedParameters);
-        }
-
         // @ts-ignore route.expand exists by now for sure
-        const { meta, layout, scripts } = route.expand;
+        const { meta, layout, scripts, context } = route.expand;
 
         expandedRoutes[url] = {
-          // @ts-expect-error This is fine
+          // TODO: Apply meta later per page
+          /*
           meta: await applyUtilities<Utility, Utilities, Context>(
             meta,
             defaultUtilities,
             { context },
           ),
+          */
+          meta: meta || {},
           layout,
           scripts,
-          context,
+          context: context || {},
           url,
         };
 
@@ -136,21 +106,6 @@ async function expandRoute(
       }));
     } else {
       console.warn("data source results are not an array");
-    }
-
-    // Take care to expand routes since they might have data source related logic etc.
-    // to execute.
-    if (recurse) {
-      const expandedRouteRoutes = await expandRoutes({
-        routes: route.routes || {},
-        dataSources,
-      });
-      allParameters = allParameters.concat(expandedRouteRoutes.allParameters);
-
-      ret = {
-        ...route,
-        routes: { ...ret.routes, ...expandedRoutes },
-      };
     }
   }
 
@@ -161,63 +116,14 @@ async function expandRoute(
       routes: route.routes || {},
       dataSources,
     });
-    allParameters = allParameters.concat(expandedRouteRoutes.allParameters);
 
     ret = {
       ...route,
-      routes: { ...ret.routes, ...expandedRouteRoutes.allRoutes },
+      routes: { ...ret.routes, ...expandedRouteRoutes },
     };
   }
 
-  if (route.dataSources) {
-    const { context, capturedParameters } = await getDataSourceContext(
-      route.dataSources,
-      dataSources,
-    );
-
-    if (capturedParameters) {
-      allParameters = allParameters.concat(capturedParameters);
-    }
-
-    return [url, { ...ret, url, context }, allParameters];
-  }
-
-  return [url, { ...ret, url }, allParameters];
-}
-
-async function getDataSourceContext(
-  dataSourceIds?: Route["dataSources"],
-  dataSources?: DataSources,
-) {
-  if (!dataSourceIds || !dataSources) {
-    return {};
-  }
-  let capturedParameters: string[] = [];
-
-  const context = Object.fromEntries(
-    await Promise.all(
-      dataSourceIds.map(async ({ name, operation, parameters }) => {
-        const dataSource = dataSources[operation];
-
-        if (!dataSource) {
-          throw new Error(`Data source ${operation} was not found!`);
-        }
-
-        capturedParameters = capturedParameters.concat(parameters as string[]);
-
-        return [
-          name,
-          await dataSource.apply(
-            undefined,
-            // @ts-expect-error This is fine
-            Array.isArray(parameters) ? parameters : [],
-          ),
-        ];
-      }),
-    ),
-  );
-
-  return { context, capturedParameters };
+  return [url, { ...ret, url }];
 }
 
 export { expandRoute, expandRoutes };
