+++ routers/getDataSourceContext.ts
@@ -1,0 +1,33 @@
+import type { DataSources, Route } from "../types.ts";
+
+async function getDataSourceContext(
+  dataSourceIds?: Route["dataSources"],
+  dataSources?: DataSources,
+) {
+  if (!dataSourceIds || !dataSources) {
+    return {};
+  }
+
+  return Object.fromEntries(
+    await Promise.all(
+      dataSourceIds.map(async ({ name, operation, parameters }) => {
+        const dataSource = dataSources[operation];
+
+        if (!dataSource) {
+          throw new Error(`Data source ${operation} was not found!`);
+        }
+
+        return [
+          name,
+          await dataSource.apply(
+            undefined,
+            // @ts-expect-error This is fine
+            Array.isArray(parameters) ? parameters : [],
+          ),
+        ];
+      }),
+    ),
+  );
+}
+
+export { getDataSourceContext };
